Correctly name downloaded Github assets with dashes or underscores in their names, share OS and architecture aliases among providers
Github assets with dashes or underscores are correctly named, by stripping the operating system and architecture that was originally used to match that asset. These components are stripped regardless of their order in the asset name.

Operating system and architecture aliases are now shared among providers. For example, `Darwin = Mac` or `amd64 = x86_64`.