Stop changing CWD, run more tests in parallel, refactor ASDF config-file functional options
The current working directory is no longer changed while extracting archive
files, or in tests. THIs helps tests to run in parallel, and archives to
be extracted concurrently in the future.

The code that searches for ASDF configuration files has been refactored
to have a cleaner functional options pattern when specifying optional
start and end (root) directories.

Finish